+++ docs/sphinx/source/whatsnew/v0.11.1.rst
@@ -16,7 +16,10 @@ Enhancements
 * Add new parameters for min/max absolute air mass to
   :py:func:`pvlib.spectrum.spectral_factor_firstsolar`.
   (:issue:`2086`, :pull:`2100`)
-
+* Restructured the pvlib/spectrum folder by breaking up the contents of
+  pvlib/spectrum/mismatch.py into pvlib/spectrum/mismatch.py,
+  pvlib/spectrum/irradiance.py, and
+  pvlib/spectrum/response.py. (:issue:`2125`, :pull:`2136`)
 * Added function for calculating wind speed at different heights,
   :py:func:`pvlib.atmosphere.windspeed_powerlaw`.
   (:issue:`2118`, :pull:`2124`)
